+++ tikione-jacocoverage-plugin/src/fr/tikione/jacocoverage/plugin/action/ProjectAntRunJ2SE.java
@@ -29,32 +29,29 @@ import org.openide.util.actions.Presenter;
  * @author Jan Lahoda (patch https://github.com/jonathanlermitage/tikione-jacocoverage/pull/3)
  */
 @ActionID(category = "Project",
-          id = "fr.tikione.jacocoverage.plugin.action.MenuAntRunProjectJ2EE")
-@ActionRegistration(displayName = "#CTL_MenuAntRunProjectJ2EE",
+          id = "fr.tikione.jacocoverage.plugin.action.ProjectAntRunJ2SE")
+@ActionRegistration(displayName = "#CTL_ProjectAntRunJ2SE",
                     lazy = false,
                     asynchronous = true,
                     surviveFocusChange = true)
 @ActionReference(path = "Projects/Actions",
-                 position = 1986,
-                 separatorBefore = 1985)
-@NbBundle.Messages("CTL_MenuAntRunProjectJ2EE=Process JaCoCoverage report...")
-public class MenuAntRunProjectJ2EE
+                 position = 1984,
+                 separatorBefore = 1983)
+@NbBundle.Messages("CTL_ProjectAntRunJ2SE=Run with JaCoCoverage")
+@SuppressWarnings("CloneableImplementsClone")
+public class ProjectAntRunJ2SE
         extends JaCoCoActionOnAnt
         implements ContextAwareAction, Presenter.Popup {
 
     private static final long serialVersionUID = 1L;
 
     @SuppressWarnings("NestedAssignment")
-    public MenuAntRunProjectJ2EE() {
+    public ProjectAntRunJ2SE() {
         super("run");
         Project project = NBUtils.getSelectedProject();
-        setEnabled(Utils.isProjectSupported(project,
-                NBProjectTypeEnum.J2EE,
-                NBProjectTypeEnum.J2EE_EAR,
-                NBProjectTypeEnum.J2EE_EJB,
-                NBProjectTypeEnum.J2EE_WEB));
+        setEnabled(Utils.isProjectSupported(project, NBProjectTypeEnum.J2SE));
         putValue(DynamicMenuContent.HIDE_WHEN_DISABLED, true);
-        putValue(Action.NAME, Bundle.CTL_MenuAntRunProjectJ2EE());
+        putValue(Action.NAME, Bundle.CTL_ProjectAntRunJ2SE());
         if (isEnabled()) { // Don't try to enable if project's type is not supported.
             FileObject prjPropsFo = project.getProjectDirectory().getFileObject("nbproject/project.properties");
             final Properties prjProps = new Properties();
@@ -62,6 +59,9 @@ public class MenuAntRunProjectJ2EE
             try {
                 if (prjPropsFo != null) {
                     prjProps.load(ins = prjPropsFo.getInputStream());
+                }
+                if (prjProps.getProperty("main.class", "").isEmpty()) {
+                    setEnabled(false);
                 }
             } catch (IOException ex) {
                 Exceptions.printStackTrace(ex);
@@ -86,7 +86,7 @@ public class MenuAntRunProjectJ2EE
 
     @Override
     public Action createContextAwareInstance(Lookup context) {
-        return new MenuAntRunProjectJ2EE();
+        return new ProjectAntRunJ2SE();
     }
 
     @Override