+++ tikione-jacocoverage-plugin/src/fr/tikione/jacocoverage/plugin/util/NBProjectTypeEnum.java
@@ -37,7 +37,6 @@ public enum NBProjectTypeEnum {
                 break;
             default:
                 qname = "n/a";
-
         }
         return qname;
     }